+++ docs/whats-new.md
@@ -1,3 +1,15 @@
+## Unreleased
+
+### Fix
+- removed `matplotlib._api` import to also support `matplotlib<3.4.0` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#465](https://github.com/Deltares/dfm_tools/pull/465)
+
+
+## 0.13.1 (2023-07-12)
+
+### Fix
+- proper fix for FM-compatible network file (1-based face_node_connectivity) by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#461](https://github.com/Deltares/dfm_tools/pull/461)
+
+
 ## 0.13.0 (2023-07-12)
 
 ### Feat